Reorganize code
- Collection method should be in collection.go
- All OpenAI compatible embedding creators can be in one file
