+++ collection.go
@@ -4,6 +4,8 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	"slices"
+	"sort"
 	"sync"
 )
 
@@ -61,6 +63,59 @@ func (c *Collection) AddConcurrently(ctx context.Context, ids []string, embeddin
 		return errors.New("concurrency must be at least 1")
 	}
 	return c.add(ctx, ids, documents, embeddings, metadatas, concurrency)
+}
+
+// Performs a nearest neighbors query on a collection specified by UUID.
+//
+//   - queryText: The text to search for.
+//   - nResults: The number of results to return. Must be > 0.
+//   - where: Conditional filtering on metadata. Optional.
+//   - whereDocument: Conditional filtering on documents. Optional.
+func (c *Collection) Query(ctx context.Context, queryText string, nResults int, where, whereDocument map[string]string) ([]Result, error) {
+	c.documentsLock.RLock()
+	defer c.documentsLock.RUnlock()
+	if len(c.documents) == 0 {
+		return nil, nil
+	}
+
+	if nResults <= 0 {
+		return nil, errors.New("nResults must be > 0")
+	}
+
+	// Validate whereDocument operators
+	for k := range whereDocument {
+		if !slices.Contains(supportedFilters, k) {
+			return nil, errors.New("unsupported operator")
+		}
+	}
+
+	// Filter docs by metadata and content
+	filteredDocs := filterDocs(c.documents, where, whereDocument)
+
+	// No need to continue if the filters got rid of all documents
+	if len(filteredDocs) == 0 {
+		return nil, nil
+	}
+
+	queryVectors, err := c.embed(ctx, queryText)
+	if err != nil {
+		return nil, fmt.Errorf("couldn't create embedding of query: %w", err)
+	}
+
+	// For the remaining documents, calculate cosine similarity.
+	res, err := calcDocSimilarity(ctx, queryVectors, filteredDocs)
+	if err != nil {
+		return nil, fmt.Errorf("couldn't calculate cosine similarity: %w", err)
+	}
+
+	// Sort by similarity
+	sort.Slice(res, func(i, j int) bool {
+		// The `less` function would usually use `<`, but we want to sort descending.
+		return res[i].Similarity > res[j].Similarity
+	})
+
+	// Return the top nResults
+	return res[:nResults], nil
 }
 
 func (c *Collection) add(ctx context.Context, ids []string, documents []string, embeddings [][]float32, metadatas []map[string]string, concurrency int) error {
